+++ extensions-core/postgresql-metadata-storage/src/main/java/org/apache/druid/metadata/storage/postgresql/PostgreSQLConnectorConfig.java
@@ -53,9 +53,7 @@ public class PostgreSQLConnectorConfig
 
   @JsonProperty
   private String sslPasswordCallback;
-  
-  @JsonProperty
-  private String dbTableSchema = "public";
+
 
   public boolean isUseSSL()
   {
@@ -107,11 +105,6 @@ public class PostgreSQLConnectorConfig
     return sslPasswordCallback;
   }
 
-  public String getDbTableSchema()
-  {
-    return dbTableSchema;
-  }
-
   @Override
   public String toString()
   {
@@ -125,7 +118,6 @@ public class PostgreSQLConnectorConfig
            ", sslRootCert='" + sslRootCert + '\'' +
            ", sslHostNameVerifier='" + sslHostNameVerifier + '\'' +
            ", sslPasswordCallback='" + sslPasswordCallback + '\'' +
-           ", dbTableSchema='" + dbTableSchema + '\'' +
-          '}';
+           '}';
   }
 }
