+++ extensions-core/postgresql-metadata-storage/src/test/java/org/apache/druid/metadata/storage/postgresql/PostgreSQLConnectorTest.java
@@ -50,7 +50,8 @@ public class PostgreSQLConnectorTest
                 null
             )
         ),
-        new PostgreSQLConnectorConfig()
+        new PostgreSQLConnectorConfig(),
+        new PostgreSQLTablesConfig()
     );
 
     Assert.assertTrue(connector.isTransientException(new SQLException("bummer, connection problem", "08DIE")));
